requirements.txt
git+https://github.com/Rapptz/discord.py.git
python-dotenv

pytest

flake8
black
